Added ws messages plugin
