+++ config.js
@@ -26,5 +26,6 @@ module.exports = [
   "shady-websockets",
   "shady-worker",
   "./plugins/live-delphi-models",
-  "./plugins/live-delphi-routes"
+  "./plugins/live-delphi-routes",
+  "./plugins/live-delphi-ws-messages"
 ];
