+++ package.json
@@ -22,7 +22,7 @@
     "nconf": "0.8.4",
     "pug": "^2.0.0-alpha6",
     "request": "2.81.0",
-    "shady-messages": "0.3.0",
+    "shady-messages": "0.3.1",
     "shady-sequelize": "0.0.1",
     "shady-websockets": "0.0.4",
     "shady-worker": "0.0.4",
