+++ plugins/live-delphi-routes/index.js
@@ -7,6 +7,7 @@
   const moment = require('moment');
   const uuid = require('uuid4');
   const config = require('nconf');
+  const util = require('util');
   const request = require('request');
   
   class Routes {
@@ -230,6 +231,43 @@
         });
     }
     
+    join(req, res) {
+      const keycloakServerUrl = config.get('keycloak:auth-server-url');
+      const keycloakRealm = config.get('keycloak:realm');
+      const keycloakUrl = util.format('%s/realms/%s/protocol/openid-connect/userinfo', keycloakServerUrl, keycloakRealm);
+      
+      request.get(keycloakUrl, {
+        'auth': {
+          'bearer': req.body.token
+        }
+      }, (authErr, response, body) => {
+        if (authErr) {
+          // TODO: Better error handling
+          this.logger.error(authErr);
+          res.status(403).send(authErr);
+        } else {
+          const reponse = JSON.parse(body);
+          const userId = reponse.sub;
+          
+          this.models.createSession(userId)
+            .then((session) => {
+              res.send({
+                sessionId: session.id
+              });
+            })
+            .catch((sessionErr) => {
+              logger.error(sessionErr);
+              res.status(500).send(sessionErr);
+            });
+        }
+      });
+    }
+    
+    getKeycloakJson(req, res) {
+      res.header('Content-Type', 'application/json');
+      res.send(config.get('keycloak'));
+    }
+    
     register(app, keycloak) {
       // Navigation
      
@@ -251,6 +289,9 @@
       app.get("/manage/queries/edit", [ keycloak.protect(), this.loggedUserMiddleware.bind(this), this.requireQueryOwner.bind(this) ], this.getEditQuery.bind(this));
       app.put("/manage/queries/edit", [ keycloak.protect(), this.loggedUserMiddleware.bind(this), this.requireQueryOwner.bind(this) ], this.putEditQuery.bind(this));
       app.delete("/manage/queries/delete", [ keycloak.protect(), this.loggedUserMiddleware.bind(this), this.requireQueryOwner.bind(this) ], this.deleteQuery.bind(this));
+      
+      app.post('/join', this.join.bind(this));
+      app.get('/keycloak.json', this.getKeycloakJson.bind(this));
     }
     
     isQueryOwner(queryId, userId) {
